+++ src/Http/Route/AdminUserClientRemoteConnectRoute.php
@@ -8,7 +8,9 @@ use Heptacom\AdminOpenAuth\Contract\OpenAuthenticationFlowInterface;
 use Shopware\Core\Framework\Api\Context\AdminApiSource;
 use Shopware\Core\Framework\Context;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -28,8 +30,14 @@ final class AdminUserClientRemoteConnectRoute extends AbstractController
         ],
         methods: ['POST']
     )]
-    public function remoteConnect(string $clientId, Context $context): Response
+    public function remoteConnect(string $clientId, Request $request, Context $context): Response
     {
+        $redirectTo = (string) $request->query->get('redirectTo') ?: null;
+
+        if ($redirectTo && !\str_starts_with($redirectTo, '/')) {
+            throw new BadRequestException('Only absolute redirect urls are allowed.');
+        }
+
         /** @var AdminApiSource $adminApiSource */
         $adminApiSource = $context->getSource();
 
@@ -37,6 +45,7 @@ final class AdminUserClientRemoteConnectRoute extends AbstractController
             'target' => $this->flow->getRedirectUrlToConnect(
                 $clientId,
                 $adminApiSource->getUserId(),
+                $redirectTo,
                 $context->scope(Context::SYSTEM_SCOPE, static fn (Context $context): Context => $context),
             ),
         ]);