+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Changed provider settings to have it's own component instead of overwriting `heptacom-admin-open-auth-client-edit-page`
 * Changed `view/administration/index-js.html.twig` to only check for relevant route matches
 * Changed `view/administration/index-js.html.twig` to allow login in `sw-inactivity-login`
+* Changed login payload to store the requested redirect url
 
 # 4.2.1